requirements.txt
absl-py==1.4.0
arrow==1.2.3
asgiref==3.6.0
astunparse==1.6.3
attrs==23.1.0
Bashutils==0.0.4
blinker==1.6.2
boto3==1.26.130
botocore==1.29.130
bravado==11.0.3
bravado-core==5.17.1
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
django-crum==0.7.9
django-layers-hr==1.11.1
experimental==0.9.7
Flask==2.3.2
flatbuffers==23.3.3
fonttools==4.39.3
fqdn==1.5.1
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
imageio==2.28.1
imutils==0.5.4
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonref==1.1.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
layers==0.1.5
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
monotonic==1.6
msgpack==1.0.5
neptune==1.1.1
neptune-tensorflow-keras==2.1.1
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pip==23.1.2
protobuf==4.23.0
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rsa==4.9
s3transfer==0.6.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sequential==1.0.0
setuptools==65.5.0
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
webcolors==1.13
websocket-client==1.5.1
Werkzeug==2.3.3
wheel==0.40.0
wrapt==1.14.1
